requirements.txt
# Django and REST Framework
Django==4.2.10
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5
markdown==3.5.2
drf-yasg==1.21.7  # API documentation
coreapi==2.3.3  # Required for DRF schema generation and docs

# Database
psycopg2-binary==2.9.9  # PostgreSQL adapter
dj-database-url==2.1.0

# Authentication
django-allauth==0.60.1
djangorestframework-simplejwt==5.3.0
django-rest-auth==0.9.5

# Environmental variables
python-dotenv==1.0.1
django-environ==0.11.2

# Background Tasks
celery==5.3.6
redis==5.0.1
django-celery-beat==2.5.0

# GIS and Data Processing
django-leaflet==0.29.0
gdal==3.7.3
numpy==1.26.3
pandas==2.2.0
scipy==1.12.0
scikit-learn==1.3.2

# Weather API integrations
requests==2.31.0
aiohttp==3.9.3

# Image processing
Pillow==10.2.0

# Development tools
black==23.12.1
flake8==7.0.0
pytest==7.4.4
pytest-django==4.8.0
coverage==7.4.1

# Deployment
gunicorn==21.2.0
whitenoise==6.6.0

# Security
django-defender==0.9.7 